+++ EarTrumpet/AppSettings.cs
@@ -132,6 +132,18 @@ namespace EarTrumpet
             set => _settings.Set("IsExpanded", value);
         }
 
+        public bool UseScrollWheelInTray
+        {
+            get => _settings.Get("UseScrollWheelInTray", true);
+            set => _settings.Set("UseScrollWheelInTray", value);
+        }
+
+        public bool UseGlobalMouseWheelHook
+        {
+            get => _settings.Get("UseGlobalMouseWheelHook", false);
+            set => _settings.Set("UseGlobalMouseWheelHook", value);
+        }
+
         public bool HasShownFirstRun
         {
             get => _settings.HasKey("hasShownFirstRun");