+++ EarTrumpet/Interop/User32.cs
@@ -352,5 +352,24 @@ namespace EarTrumpet.Interop
 
         [DllImport("user32.dll", PreserveSig = true)]
         public static extern IntPtr SendMessage(IntPtr hWnd, int wMsg, IntPtr wParam, IntPtr lParam);
+
+        public delegate int HookProc(int nCode, IntPtr wParam, IntPtr lParam);
+
+        [DllImport("user32.dll", PreserveSig = true)]
+        public static extern int SetWindowsHookEx(
+            int idHook,
+            HookProc lpfn,
+            IntPtr hInstance,
+            int threadId);
+
+        [DllImport("user32.dll", PreserveSig = true)]
+        public static extern bool UnhookWindowsHookEx(int idHook);
+
+        [DllImport("user32.dll", PreserveSig = true)]
+        public static extern int CallNextHookEx(
+            int idHook,
+            int nCode,
+            IntPtr wParam,
+            IntPtr lParam);
     }
 }